Add method to query Lieutenant API
This commit adds a method to query the Lieutenant API. This method is
quite opinionated and expects the Lieutenant API token to be present in
environment variable `LIEUTENANT_API_TOKEN`.

The method is Promise-based and makes use of our `httpsQuery` method and
transforms the API responses to `ClusterInfo` objects for further use in
the Commodore manager.