Prefix `extra.yaml` with name of file we're renovating
Previously we generated a file named `extra.yaml` for each file that's
getting renovated. This was fine because the contents of `extra.yaml`
were static for each repo that's getting renovated.

However, now we fetch dynamic information from Lieutenant when
renovating tenant repos, so we want to ensure that each file is
renovated with the data that Lieutenant returned for that file.

Note that this change doesn't guarantee that we use the right extra data
if there's multiple files with the same base name in a tenant repo.