Expose Lieutenant token config only via environment variable
Instead of exposing a `lieutenantToken` config option which defaults to
environment variable `LIEUTENANT_API_TOKEN`, we expose a config option
which allows specifying the environment variable from which the token
should be read.

With the previous approach, there was no way for users to securely
specify different Lieutenant tokens for different repositories, except
for providing the token in plain text in the repository's `renovate.json`.

With the new approach, users can provide all the required Lieutenant
secrets as environment variables in the Renovate runner execution
environment, and users can select an appropriate token for their
configured Lieutenant instance.

Note: even though we don't expose the `lieutenantToken` option anymore,
we internally still set the field on the config object, so the commodore
manager implementation can still read the token from the provided config
object.

To achieve this, we need to patch Renovate's `getRepositoryConfig`
method. However, since that method is in the same module as it's
defined, patching it directly doesn't inject our patch where we need it.
Instead we patch `filterConfig` which is called by `getRepositoryConfig`
to set `config.lieutenantToken` to the value of the environment variable
specified with `config.lieutenanttokenEnvVar`, if that environment
variable is set.