Read Lieutenant API token from environment variable in Commodore manager
The approach with patching the Renovate config loader didn't work as
expected without also exposing the field `lieutenantToken` to users
directly.

To ensure that users don't configure tokens in plain text we adjust the
implementation to read the token from the environment variable in
`extractPackageFile` and set field `config.lieutenantToken` there.