Only keep `syn-commodore` in requirements.txt
This enables us to let Renovate update the Commodore version used for
commodore-renovate. We do this although it's not really best practice,
because we otherwise have issues with Renovate updating transitive
dependencies to versions which are incompatible with the version of that
dependency which is required by Kapitan.

Since we don't rebuild the image frequently, not having all Commodore
dependencies pinned shouldn't be a big issue, we've been running a
requirements.txt like this for the Commodore component template without
any issues so far.