Fix `fileMatch` for commodore-helm manager
We need to anchor the beginning of the `fileMatch` commodore-helm regex
`class/[^.].ya?ml$` so that we don't accidentally match any files in
golden tests whose name or path contains the fragment `class/`, such as
component-rook-ceph's `rbd-extra-storageclass` test case.

Unfortunately, writing a test for this is tricky because we're
configuring Renovate internals which we can't easily exercise in our
test cases.